Structure the app files to use the ducks pattern for Redux files
