+++ src/App.js
@@ -1,31 +1,12 @@
-import { Route, Routes, NavLink } from 'react-router-dom';
+import { Route, Routes } from 'react-router-dom';
 import './App.css';
-import Categories from './pages/Categories';
-import Books from './pages/Books';
-import userIcon from './images/userIcon.png';
-
-const links = [
-  { path: '/', text: 'Books' },
-  { path: 'categories', text: 'Categories' },
-];
-
-const titleStyles = { color: '#0290ff', fontSize: '1.875rem', fontFamily: 'Montserrat-Bold' };
+import Categories from './components/pages/Categories';
+import Books from './components/pages/Books';
+import Navbar from './components/Navbar';
 
 const App = () => (
   <div className="main-container">
-    <nav>
-      <h1 style={titleStyles}>Bookstore CMS</h1>
-      <div className="nav_top">
-        <ul>
-          {links.map((link) => (
-            <li key={link.text}>
-              <NavLink to={link.path}>{link.text}</NavLink>
-            </li>
-          ))}
-        </ul>
-        <img src={userIcon} alt="user icon" />
-      </div>
-    </nav>
+    <Navbar />
     <section>
       <Routes>
         <Route path="/" element={<Books />} />
